+++ packages/ritext/src/lib/extensions/Strike.tsx
@@ -1,16 +1,16 @@
-import TiptapBold from "@tiptap/extension-bold";
+import TiptapStrike from "@tiptap/extension-strike";
 import { Mark } from "@tiptap/react";
-import BoldIcon from "../icon/BoldIcon";
+import StrikeIcon from "../icon/StrikeIcon";
 
 //Components
 import ButtonComponent from "../components/ButtonComponent";
 import { ExtButtonOptions } from "../types/tiptap-ext.type";
 
-export const Bold: Mark<ExtButtonOptions, any> = TiptapBold.extend<ExtButtonOptions>({
+export const Strike: Mark<ExtButtonOptions, any> = TiptapStrike.extend<ExtButtonOptions>({
     addOptions() {
         return {
             ...this.parent?.(),
-            button: ({ editor, options, buttonClassName }) => {
+            button: ({ options, editor, buttonClassName }) => {
                 return (
                     <ButtonComponent
                         className={options.className}
@@ -20,11 +20,11 @@ export const Bold: Mark<ExtButtonOptions, any> = TiptapBold.extend<ExtButtonOpti
                         tooltip={options.tooltip}
                         tooltipClassName={options.tooltipClassName}
                         tooltipPlacement={options.tooltipPlacement}
-                        _internalIcon={<BoldIcon />}
-                        _extName="bold"
-                        _onToggle={() => editor.chain().focus().toggleBold().run()}
-                        _interShortcut="⌘ + B"
-                        _tooltipContent="Bold"
+                        _internalIcon={<StrikeIcon />}
+                        _extName="strike"
+                        _onToggle={() => editor.chain().focus().toggleStrike().run()}
+                        _interShortcut="⌘ ⇧ S"
+                        _tooltipContent="Strike"
                         _buttonClassName={buttonClassName}
                     />
                 )