+++ packages/ritext/src/lib/types/tiptap-ext.type.ts
@@ -1,0 +1,14 @@
+import { ReactNode, CSSProperties } from "react";
+import { Placement } from "@floating-ui/react";
+import { Editor } from "@tiptap/react";
+
+export type ExtButtonOptions = {
+    className?: string;
+    activeClassName?: string;
+    icon?: ReactNode,
+    style?: CSSProperties,
+    tooltip?: boolean | string;
+    tooltipClassName?: string;
+    tooltipPlacement?: Placement;
+    button: (args: { options: ExtButtonOptions, editor: Editor, buttonClassName: string }) => ReactNode;
+}