+++ CHANGELOG.md
@@ -30,3 +30,23 @@
 ### Patch Changes
 
 - Added new features
+
+---
+
+### ritext
+
+## 1.0.11
+
+### Patch Changes
+
+- Added italic, underline and strike component
+
+---
+
+### ritext
+
+## 1.0.12
+
+### Patch Changes
+
+- Added italic, underline and strike extension